requirements.txt
PyNaCl>=1.5.0
cryptography>=42.0.0
pycryptodome>=3.20.0
